requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
torch
opencv-python-headless
pillow
ultralytics
pandas
tqdm
seaborn
